requirements.txt
Flask==3.1.2
flask-cors==6.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1

bcrypt==4.1.3
python-dotenv==1.0.1

google-genai==1.73.1

pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.5

fairlearn==0.13.0
scikit-learn==1.8.0

requests==2.31.0

pytest==8.2.0